data-pipeline/requirements.txt
requests
pandas
supabase
SPARQLWrapper
python-dotenv
pydantic